Added MockupDataGenerator. Wired with all components that use mockup data.
